# docs/swagger.yaml
basePath: /api/v1
definitions:
  apperr.Code:
    enum:
    - core/bad_request
    - core/unauthorized
    - core/forbidden
    - core/internal_error
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeInternal
  apperr.Field:
    enum:
    - request
    type: string
    x-enum-varnames:
    - FieldRequest
  apperr.Rule:
    enum:
    - required
    - too_long
    - cycle
    - max_hierarchy_depth
    - invalid_format
    - too_short
    - duplicate
    - mismatch
    - forbidden
    - invalid_state
    - not_found
    type: string
    x-enum-varnames:
    - RuleRequired
    - RuleTooLong
    - RuleCycle
    - RuleMaxHierarchy
    - RuleInvalidFormat
    - RuleTooShort
    - RuleDuplicate
    - RuleMismatch
    - RuleForbidden
    - RuleInvalidState
    - RuleNotFound
  apperr.Violation:
    properties:
      field:
        $ref: '#/definitions/apperr.Field'
      params:
        additionalProperties: {}
        type: object
      rule:
        $ref: '#/definitions/apperr.Rule'
    type: object
  apperr.appError:
    properties:
      code:
        $ref: '#/definitions/apperr.Code'
      message:
        description: Message for user
        type: string
      violations:
        items:
          $ref: '#/definitions/apperr.Violation'
        type: array
    type: object
...
  entity.Entity:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      current_version:
        type: integer
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      type:
        $ref: '#/definitions/entity.Type'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  entity.Node:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.Node'
        type: array
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      type:
        $ref: '#/definitions/entity.Type'
    type: object
  entity.Type:
    enum:
    - article
    - department
    type: string
    x-enum-varnames:
    - TypeArticle
    - TypeDepartment
...
  http.CreateEntityResp:
    properties:
      id:
        type: string
    type: object
...
  http.UpdateEntityInput:
    properties:
      content:
        type: string
      is_draft:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
    type: object
...
  usecase.CreateEntityCmd:
    properties:
      content:
        type: string
      is_draft:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      type:
        $ref: '#/definitions/entity.Type'
    type: object
...
info:
  contact: {}
  description: Demo wiki backend in Go
  title: EasyGoDocs API
  version: "1.0"
paths:
  /entities:
    get:
      description: Returns the hierarchical tree of all permitted entities.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Node'
            type: array
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Get full entity tree
      tags:
      - entities
    post:
      consumes:
      - application/json
      description: Creates a new entity. Requires write permission for the parent
        entity. if root entity, requires admin role.
      parameters:
      - description: Create entity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateEntityCmd'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.CreateEntityResp'
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Create entity
      tags:
      - entities
  /entities/{entity_id}:
    delete:
      description: Deletes an entity by ID. Requires write permission for the entity.
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Delete entity
      tags:
      - entities
    get:
      description: Returns a single entity by its ID. Requires read permission.
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Entity'
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Get entity by ID
      tags:
      - entities
    put:
      consumes:
      - application/json
      description: Updates an existing entity. Requires write permission. If changes
        parent, requires write permission for the new and old parents as well.
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      - description: Update entity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateEntityInput'
      responses:
        "204":
          description: No Content
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Update entity
      tags:
      - entities
  /entities/{entity_id}/versions:
    get:
      description: Returns list of all versions for an entity. Requires read permission.
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Entity'
            type: array
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: List entity versions
      tags:
      - entities
  /entities/{entity_id}/versions/{version}:
    get:
      description: Returns a specific version of an entity. Requires read permission.
      parameters:
      - description: Entity ID
        in: path
        name: entity_id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Entity'
        default:
          description: Error
          schema:
            $ref: '#/definitions/apperr.appError'
      security:
      - BearerAuth: []
      summary: Get specific entity version
      tags:
      - entities
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"